+++ projects/demo-maps/src/app/route-components/route-example-layer-style/route-map6.component.ts
@@ -9,7 +9,7 @@ import { RasterFeatureInfoComponent } from '../../components/raster-feature-info
 import { ClarityIcons, layersIcon, worldIcon } from '@cds/core/icon';
 import { MapOlComponent } from '@dlr-eoc/map-ol';
 import { ClrVerticalNavModule, ClrStandaloneCdkTrapFocus, ClrNavigationModule, ClrIconModule } from '@clr/angular';
-import { LayerControlComponent, BaseLayerControlComponent } from '@dlr-eoc/layer-control';
+import { LayerControlComponent, BaseLayerControlComponent } from '@dlr-eoc/ngx-ukis-ui-clarity';
 ClarityIcons.addIcons(...[layersIcon, worldIcon]);
 
 @Component({