+++ projects/layer-control/src/public-api.ts
@@ -1,8 +1,0 @@
-/*
- * Public API Surface of layer-control
- */
-
-export * from './lib/layer-control/layer-control.component';
-export * from './lib/base-layer-control/base-layer-control.component';
-export * from './lib/layerentry/layerentry.component';
-export * from './lib/layerentry-group/layerentry-group.component';