+++ projects/user-info/src/public-api.ts
@@ -1,8 +1,0 @@
-/*
- * Public API Surface of user-info
- */
-
-export * from './lib/user.service';
-export * from './lib/login/login.component';
-export * from './lib/register/register.component';
-export * from './lib/user-details/user-details.component';